+++ gui/renderers/RayTracer.cpp
@@ -9,32 +9,12 @@
 
 NAMESPACE_SPH_BEGIN
 
-inline auto seeder() {
-    return [seed = 1337]() mutable { return seed++; };
-}
-
-RayMarcher::ThreadData::ThreadData(const int seed)
-    : rng(seed) {}
-
 RayMarcher::RayMarcher(SharedPtr<IScheduler> scheduler, const GuiSettings& settings)
-    : scheduler(scheduler)
-    , threadData(*scheduler, seeder()) {
+    : IRaytracer(scheduler, settings) {
     kernel = CubicSpline<3>();
     fixed.dirToSun = getNormalized(settings.get<Vector>(GuiSettingsId::SURFACE_SUN_POSITION));
     fixed.brdf = Factory::getBrdf(settings);
-    fixed.colorMap = Factory::getColorMap(settings);
-    fixed.subsampling = settings.get<int>(GuiSettingsId::RAYTRACE_SUBSAMPLING);
-    fixed.iterationLimit = settings.get<int>(GuiSettingsId::RAYTRACE_ITERATION_LIMIT);
-
-    fixed.enviro.color = settings.get<Rgba>(GuiSettingsId::BACKGROUND_COLOR);
-    std::string hdriPath = settings.get<std::string>(GuiSettingsId::RAYTRACE_HDRI);
-    if (!hdriPath.empty()) {
-        fixed.enviro.hdri = Texture(Path(hdriPath), TextureFiltering::BILINEAR);
-    }
-
     fixed.renderSpheres = settings.get<bool>(GuiSettingsId::RAYTRACE_SPHERES);
-
-    shouldContinue = true;
 }
 
 RayMarcher::~RayMarcher() = default;
@@ -133,106 +113,27 @@ void RayMarcher::initialize(const Storage& storage,
     finder->build(*scheduler, cached.r);
 
     for (ThreadData& data : threadData) {
-        data.previousIdx = Size(-1);
+        MarchData march;
+        march.previousIdx = Size(-1);
+        data.data = std::move(march);
     }
 
     shouldContinue = true;
 }
 
-void RayMarcher::render(const RenderParams& params, Statistics& UNUSED(stats), IRenderOutput& output) const {
-    shouldContinue = true;
-    SPH_ASSERT(finder && bvh.getBoundingBox().volume() > 0._f);
-    FrameBuffer fb(params.size);
-    for (Size iteration = 0; iteration < fixed.iterationLimit && shouldContinue; ++iteration) {
-        this->refine(params, iteration, fb);
+Rgba RayMarcher::shade(const RenderParams& params, const CameraRay& cameraRay, ThreadData& data) const {
+    const Vector dir = getNormalized(cameraRay.target - cameraRay.origin);
+    const Ray ray(cameraRay.origin, dir);
 
-        const bool isFinal = (iteration == fixed.iterationLimit - 1);
-        if (fixed.colorMap) {
-            Bitmap<Rgba> bitmap = fixed.colorMap->map(fb.getBitmap());
-            output.update(bitmap, {}, isFinal);
-        } else {
-            output.update(fb.getBitmap(), {}, isFinal);
-        }
-    }
-}
-
-INLINE float sampleTent(const float x) {
-    if (x < 0.5f) {
-        return sqrt(2.f * x) - 1.f;
+    MarchData& march(data.data);
+    if (Optional<Vector> hit = this->intersect(march, ray, params.surface.level, false)) {
+        return this->getSurfaceColor(march, params, march.previousIdx, hit.value(), ray.direction());
     } else {
-        return 1.f - sqrt(1.f - 2.f * (x - 0.5f));
+        return this->getEnviroColor(cameraRay);
     }
 }
 
-INLINE Coords sampleTent2d(const Size level, const float halfWidth, UniformRng& rng) {
-    if (level == 1) {
-        const float x = 0.5f + sampleTent(float(rng())) * halfWidth;
-        const float y = 0.5f + sampleTent(float(rng())) * halfWidth;
-        return Coords(x, y);
-    } else {
-        // center of the pixel
-        return Coords(0.5f, 0.5f);
-    }
-}
-
-void RayMarcher::refine(const RenderParams& params, const Size iteration, FrameBuffer& fb) const {
-    MEASURE_SCOPE("Rendering frame");
-    const Size level = 1 << max(int(fixed.subsampling) - int(iteration), 0);
-    Pixel actSize;
-    actSize.x = params.size.x / level + sgn(params.size.x % level);
-    actSize.y = params.size.y / level + sgn(params.size.y % level);
-    Bitmap<Rgba> bitmap(actSize);
-
-    const bool first = (iteration == 0);
-    parallelFor(*scheduler,
-        threadData,
-        0,
-        Size(bitmap.size().y),
-        1,
-        [this, &bitmap, &params, level, first](Size y, ThreadData& data) {
-            if (!shouldContinue && !first) {
-                return;
-            }
-            for (Size x = 0; x < Size(bitmap.size().x); ++x) {
-                const Coords pixel = Coords(x * level, y * level) +
-                                     sampleTent2d(level, params.surface.filterWidth / 2.f, data.rng);
-                const Optional<CameraRay> cameraRay = params.camera->unproject(pixel);
-                if (!cameraRay) {
-                    bitmap[Pixel(x, y)] = Rgba::black();
-                    continue;
-                }
-
-                const Vector dir = getNormalized(cameraRay->target - cameraRay->origin);
-                const Ray ray(cameraRay->origin, dir);
-
-                Rgba accumulatedColor = Rgba::transparent();
-                if (Optional<Vector> hit = this->intersect(data, ray, params.surface.level, false)) {
-                    accumulatedColor =
-                        this->shade(data, params, data.previousIdx, hit.value(), ray.direction());
-                } else {
-                    accumulatedColor = this->getEnviroColor(ray);
-                }
-                bitmap[Pixel(x, y)] = accumulatedColor;
-            }
-        });
-
-    if (!shouldContinue && !first) {
-        return;
-    }
-    if (level == 1) {
-        fb.accumulate(bitmap);
-    } else {
-        Bitmap<Rgba> full(params.size);
-        for (Size y = 0; y < Size(full.size().y); ++y) {
-            for (Size x = 0; x < Size(full.size().x); ++x) {
-                full[Pixel(x, y)] = bitmap[Pixel(x / level, y / level)];
-            }
-        }
-        fb.override(std::move(full));
-    }
-}
-
-ArrayView<const Size> RayMarcher::getNeighbourList(ThreadData& data, const Size index) const {
+ArrayView<const Size> RayMarcher::getNeighbourList(MarchData& data, const Size index) const {
     // look for neighbours only if the intersected particle differs from the previous one
     if (index != data.previousIdx) {
         Array<NeighbourRecord> neighs;
@@ -252,7 +153,7 @@ ArrayView<const Size> RayMarcher::getNeighbourList(ThreadData& data, const Size 
     return data.neighs;
 }
 
-Optional<Vector> RayMarcher::intersect(ThreadData& data,
+Optional<Vector> RayMarcher::intersect(MarchData& data,
     const Ray& ray,
     const Float surfaceLevel,
     const bool occlusion) const {
@@ -273,7 +174,7 @@ Optional<Vector> RayMarcher::intersect(ThreadData& data,
     return NOTHING;
 }
 
-Optional<Vector> RayMarcher::getSurfaceHit(ThreadData& data,
+Optional<Vector> RayMarcher::getSurfaceHit(MarchData& data,
     const IntersectContext& context,
     bool occlusion) const {
     if (fixed.renderSpheres) {
@@ -325,7 +226,7 @@ Optional<Vector> RayMarcher::getSurfaceHit(ThreadData& data,
     }
 }
 
-Rgba RayMarcher::shade(ThreadData& data,
+Rgba RayMarcher::getSurfaceColor(MarchData& data,
     const RenderParams& params,
     const Size index,
     const Vector& hit,
@@ -453,18 +354,6 @@ Vector RayMarcher::evalUvws(ArrayView<const Size> neighs, const Vector& pos1) co
     } else {
         SPH_ASSERT(weightSum != 0._f);
         return uvws / weightSum;
-    }
-}
-
-Rgba RayMarcher::getEnviroColor(const Ray& ray) const {
-    if (fixed.enviro.hdri.empty()) {
-        return fixed.enviro.color;
-    } else {
-        const Vector dir = ray.direction();
-        /// \todo deduplicate with setupUvws
-        const SphericalCoords spherical = cartensianToSpherical(Vector(dir[X], dir[Z], dir[Y]));
-        const Vector uvw(spherical.phi / (2._f * PI) + 0.5_f, spherical.theta / PI, 0._f);
-        return fixed.enviro.hdri.eval(uvw);
     }
 }